started with basic websocket
Started a basic websocket. Turns out webSocket is acompletly different
protocol working over TCP to have handshake and send data in its own
frame. One alternative is to use Kazing's Websocket gateway to redirect
traffic coming to websocket to our server running UDP and back. In short
a kinda proxy.

Good news is WebSockets is working on all browsers (including mobile) so
it wont have a problem and the latency is near to that of UDP as the
over head is extremly less.